requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
scipy==1.11.3
requests==2.31.0